[REF] Rename module, clean code and refactory
