+++ pos_payment_term/static/src/js/models.js
@@ -1,22 +1,10 @@
-/******************************************************************************
-*    Point Of Sale - Payment Terms for POS Odoo
-*    Copyright (C) 2016 KMEE INFORMATICA LTDA (http://www.kmee.com.br)
-*    @author Luiz Felipe do Divino <luiz.divino@kmee.com.br>
-*
-*    This program is free software: you can redistribute it and/or modify
-*    it under the terms of the GNU Affero General Public License as
-*    published by the Free Software Foundation, either version 3 of the
-*    License, or (at your option) any later version.
-*    This program is distributed in the hope that it will be useful,
-*    but WITHOUT ANY WARRANTY; without even the implied warranty of
-*    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-*    GNU Affero General Public License for more details.
-*    You should have received a copy of the GNU Affero General Public License
-*    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*
+/*****************************************************************************
+* © 2016 KMEE INFORMATICA LTDA (<http://kmee.com.br>)
+*    Luiz Felipe do Divino <luiz.divino@kmee.com.br>
+* License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 ******************************************************************************/
 
-function pos_payment_terms_models(instance, module){
+function pos_payment_term_models(instance, module){
 
     var PosModelParent = module.PosModel;
     module.PosModel = module.PosModel.extend({
