+++ pos_payment_term/static/src/xml/pos.xml
