+++ pos_payment_term/wizard/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+# © 2016 KMEE INFORMATICA LTDA (<http://kmee.com.br>). - Luiz Felipe do Divino <luiz.divino@kmee.com.br>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from . import pos_payment
