+++ pos_payment_term/wizard/pos_payment.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+        <record id="pos_payment_term_view_pos_payment" model="ir.ui.view">
+            <field name="name">Pay Order</field>
+            <field name="model">pos.make.payment</field>
+            <field name="inherit_id"
+                   ref="point_of_sale.view_pos_payment" />
+            <field name="arch" type="xml">
+                <field name="journal_id" position="after">
+                    <field name="amount" />
+                </field>
+            </field>
+        </record>
+    </data>
+</openerp>
+
