--- pos_payment_terms/__init__.py
@@ -1,2 +1,0 @@
-from . import models
-from . import wizard
