--- pos_payment_terms/__openerp__.py
@@ -1,38 +1,0 @@
-###############################################################################
-#     Point Of Sale - Payment Terms for POS Odoo
-#    Copyright (C) 2016 KMEE INFORMATICA LTDA (http://www.kmee.com.br)
-#    @author Luiz Felipe do Divino <luiz.divino@kmee.com.br>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-###############################################################################
-
-{
-    "name": "Point of Sale - Payment Terms",
-    "version": "8.0.1.0.0",
-    "author": "KMEE INFORMATICA LTDA, "
-              "Odoo Community Association (OCA)",
-    'website': 'http://odoo-brasil.org',
-    "license": "AGPL-3",
-    "category": "Point Of Sale",
-    "depends": [
-        'point_of_sale',
-        'pos_pricelist',
-        'account',
-    ],
-    'data': [
-        "views/pos_template.xml",
-    ],
-    "qweb": [
-        'static/src/xml/pos.xml',
-    ],
-    "installable": True,
-}
