--- pos_payment_terms/models/__init__.py
@@ -1,1 +1,0 @@
-from . import point_of_sale
