--- pos_payment_terms/wizard/__init__.py
@@ -1,1 +1,0 @@
-from . import pos_payment
