--- pos_payment_terms/wizard/pos_payment.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record id="view_pos_payment" model="ir.ui.view">
-            <field name="name">Pay Order</field>
-            <field name="model">pos.make.payment</field>
-            <field name="arch" type="xml">
-            <form string="Pay Order">
-                <group>
-                    <field name="journal_id" />
-                    <field name="amount" />
-                    <field name="payment_name"/>
-                </group>
-                <footer>
-                    <button name="check" string="Make Payment" type="object" class="oe_highlight"/>
-                    or
-                    <button special="cancel" string="Cancel" class="oe_link"/>
-                </footer>
-            </form>
-            </field>
-        </record>
-        <record id="action_pos_payment" model="ir.actions.act_window">
-            <field name="name">Payment</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">pos.make.payment</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">form</field>
-            <field name="target">new</field>
-        </record>
-    </data>
-</openerp>
-
